Overhaul bootloader and kernel improvements.
PATCH NOTES

<1> (NEW) New Bootloader (with FAT FS for Boot Disk)
<2> (NEW) AC97 Audio Support (partial)
<3> (NEW) VFS File System
<4> (IMPROVED) RTL8139 Networking Support
<5> (IMPROVED) Bochs VBE Extension Support

DEVELOPMENTS

Further information about the following will be released at a later date.

<+> (DEVELOPING) New Tasking System
<+> (DEVELOPING) Support for System Calls
<+> (DEVELOPING) Support for ELF Executables

1. New Bootloader

The previous bootloader was extremely primitive.

- It did not read the memory map.
- It used memory addresses rather than files (no file system).
- It was just disorganized in general.

The new bootloader takes place in three stages: Stage 0, Stage 1, and Stage 2.

All three stages together take up the first track (18 sectors) of the disk.

Stage 0 is first. It contains the FAT12 header and is just a menu with three options.
- Continue - Reads Stage 1 into memory and jumps to it.
- Shutdown - Will cause the system to shut down.
- About    - Changes the color of the text on the screen (I ran out of bytes to add anything meaningful).

Stage 1 will read the memory map and disk information before enabling protected mode, before reading and jumping to Stage 2.

Stage 2 will read the disk and find a file called "KERNEL.BIN", which, if found, will be read and written to a memory address which it will then jump to.

Using this approach, we can also have files other than the kernel on the disk, such as audio files (spoiler alert).

2. AC97 Audio Support

A personal goal of mine in this project was to have high-definition audio playback support.

After a lot of work, I finally did it — sort of. I was able to add basic support for the AC97 audio card.
I say sort of because it can't play audio files larger than 1 second at 44,100 Hz.
But for short sound effects, it works fine.

Only .wav files are currently supported because they are, for the most part, raw PCM.

Further work will be done on this in the future.

3. VFS File System

By adding a Virtual File System, the file system is now a lot more organized.
It's still not perfect, but it is much better than just having a disorganized RAMFS.
It also (in theory) allows us to add support for real file systems in the future.

4. RTL8139 Networking Improvements

Reorganized much of the networking code and made certain optimizations to the driver.
I've also been experimenting with IP, but that's on hold for now.

5. Bochs VBE Extension Improvements

Previously, Bochs VBE "worked", but it was very slow due to using banking instead of a linear framebuffer.

After some research, I figured out how to use the linear framebuffer rather than banking, and performance massively improved.

There are still some performance issues which will still need to be figured out, but this was a great start.